New Event Form Completed
Functionality to navigate from homepage to a form to add new events, submit new event data and see that event populate in the list on the homepage is finished. The ViewModel keeps track of input values and resets them once an event is submitted. Next steps include adding error checking and fixing the prompts and format of the date fields.