Completed basics of NewEventForm
The NewEventForm has been completed, with basics of error handling as well as an attempt at recreating a non-experimental date picker. RegEx is used to confirm date format, other than that strings need to just not be empty since they're mostly for informative display. Next steps involve creating the TopAppBar and the Tabs that will help the user navigate between leagues, tournaments and events on the homepage.